feat(card): add legacy dark theme tokens
